requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
attrs==22.1.0
awscli==1.25.86
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
bokeh==2.4.3
botocore==1.27.85
bs4==0.0.1
cattrs==22.1.0
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
comet-ml==3.31.19
configobj==5.0.6
contourpy==1.0.5
cookiecutter==2.1.1
coverage==6.5.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.16
dulwich==0.20.50
entrypoints==0.4
everett==3.1.0
exceptiongroup==1.0.0rc9
executing==1.1.0
fastjsonschema==2.16.2
flake8==5.0.4
fonttools==4.37.4
idna==3.4
imageio==2.22.2
imagesize==1.4.1
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.4.12
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
panel==0.14.0
param==1.12.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pyct==0.4.8
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-cache==0.9.6
requests-toolbelt==0.10.1
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.11.0
simplejson==3.18.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
-e git+https://github.com/AlainFidahoussen/IFT6758-A2022-G08.git@7c0930c987ba10003d0288a2a2143d6cc423e6f0#egg=src
stack-data==0.5.1
tenacity==8.1.0
terminado==0.16.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
url-normalize==1.4.3
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
widgetsnbextension==4.0.3
wrapt==1.14.1
wurlitzer==3.0.2
xgboost==1.7.1
zipp==3.8.1
